refactor: move v2 iiif builders to iiif/v2 folder
