+++ libs/ngx-mime/src/lib/core/builders/iiif/v2/canvas.builder.ts
@@ -1,4 +1,4 @@
-import { Canvas } from '../models/manifest';
+import { Canvas } from '../../../models/manifest';
 import { BuilderUtils } from './builder-utils';
 import { ImagesBuilder } from './images.builder';
