+++ libs/ngx-mime/src/lib/core/builders/iiif/v2/manifest.builder.spec.ts
@@ -1,5 +1,5 @@
-import { testManifest } from '../../test/testManifest';
-import { Canvas, Images, Manifest, Sequence } from '../models/manifest';
+import { testManifest } from '../../../../test/testManifest';
+import { Canvas, Images, Manifest, Sequence } from '../../../models/manifest';
 import { ManifestBuilder } from './manifest.builder';
 
 describe('ManifestBuilder', () => {
@@ -51,7 +51,9 @@ describe('ManifestBuilder', () => {
       const firstSequence = manifest.sequences[0];
       if (firstSequence.canvases && firstSequence.canvases.length > 0) {
         const firstCanvas = firstSequence.canvases[0];
-        expect(firstCanvas.altoUrl).toBeDefined('First canvas should have altoUrl');
+        expect(firstCanvas.altoUrl).toBeDefined(
+          'First canvas should have altoUrl'
+        );
       }
     }
   });
