+++ libs/ngx-mime/src/lib/core/builders/iiif/v2/metadata.builder.ts
@@ -1,4 +1,4 @@
-import { Metadata } from '../models/manifest';
+import { Metadata } from '../../../models/manifest';
 
 export class MetadataBuilder {
   constructor(private metadatas: any) {}
