+++ libs/ngx-mime/src/lib/core/builders/iiif/v2/resource.builder.ts
@@ -1,4 +1,4 @@
-import { Resource } from '../models/manifest';
+import { Resource } from '../../../models/manifest';
 import { BuilderUtils } from './builder-utils';
 import { ServiceBuilder } from './service.builder';
