+++ libs/ngx-mime/src/lib/core/builders/iiif/v2/sizes.builder.ts
@@ -1,4 +1,4 @@
-import { Size } from '../models/manifest';
+import { Size } from '../../../models/manifest';
 
 export class SizesBuilder {
   constructor(private sizes: any[]) {}
