+++ libs/ngx-mime/src/lib/core/builders/iiif/v2/structure.builder.ts
@@ -1,4 +1,4 @@
-import { Structure, Sequence } from '../models/manifest';
+import { Sequence, Structure } from '../../../models/manifest';
 import { BuilderUtils } from './builder-utils';
 
 export class StructureBuilder {
@@ -18,7 +18,7 @@ export class StructureBuilder {
             canvasIndex: BuilderUtils.findCanvasIndex(
               structure.canvases,
               this.sequences
-            )
+            ),
           })
         );
       }
