+++ libs/ngx-mime/src/lib/core/iiif-content-search-service/iiif-content-search.service.ts
@@ -2,8 +2,8 @@ import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable, Subject, throwError } from 'rxjs';
 import { distinctUntilChanged, finalize, take } from 'rxjs/operators';
+import { SearchResultBuilder } from '../builders/iiif/search-result.builder';
 import { MimeViewerConfig } from '../mime-viewer-config';
-import { SearchResultBuilder } from './../builders/search-result.builder';
 import { Hit } from './../models/hit';
 import { IiifSearchResult } from './../models/iiif-search-result';
 import { Manifest } from './../models/manifest';
@@ -11,9 +11,8 @@ import { SearchResult } from './../models/search-result';
 
 @Injectable()
 export class IiifContentSearchService {
-  protected _currentSearchResult: Subject<SearchResult> = new BehaviorSubject<SearchResult>(
-    new SearchResult({})
-  );
+  protected _currentSearchResult: Subject<SearchResult> =
+    new BehaviorSubject<SearchResult>(new SearchResult({}));
   protected _searching = new BehaviorSubject<boolean>(false);
   protected _currentQ = new BehaviorSubject<string>('');
   protected _selected = new BehaviorSubject<Hit | null>(null);
