+++ libs/ngx-mime/src/lib/core/iiif-manifest-service/iiif-manifest-service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { distinctUntilChanged, finalize, take } from 'rxjs/operators';
-import { ManifestBuilder } from '../builders/manifest.builder';
+import { ManifestBuilder } from '../builders/iiif/v2/manifest.builder';
 import { MimeViewerIntl } from '../intl/viewer-intl';
 import { Manifest } from '../models/manifest';
 import { SpinnerService } from '../spinner-service/spinner.service';
@@ -30,7 +30,7 @@ export class IiifManifestService {
     return new Observable((observer) => {
       if (manifestUri.length === 0) {
         this._errorMessage.next(this.intl.manifestUriMissingLabel);
-        observer.next(false)
+        observer.next(false);
       } else {
         this.spinnerService.show();
         this.http
@@ -44,15 +44,15 @@ export class IiifManifestService {
               const manifest = this.extractData(response);
               if (this.isManifestValid(manifest)) {
                 this._currentManifest.next(manifest);
-                observer.next(true)
+                observer.next(true);
               } else {
                 this._errorMessage.next(this.intl.manifestNotValidLabel);
-                observer.next(false)
+                observer.next(false);
               }
             },
             (err: HttpErrorResponse) => {
               this._errorMessage.next(this.handleError(err));
-              observer.next(false)
+              observer.next(false);
             }
           );
       }
